Make parameter required, since it has a default (#113)
* Make parameter required, since it has a default

Some WDL implementations (such as `miniwdl`) flag the use of an optional type (e.g. `String?`) in an equality check as an invalid type conversion. This seems to deviate from the spec. However in this case, `strandedness` has a default value specified. So there shouldn't be any need to make it optional as well. This just makes it a required parameter to alleviate errors when running with `miniwdl`. It shouldn't have any effect on the workflow itself.

* Removing additional unneeded optional flags

Both `read_length` and `insert_size` have default values. Therefore the optional flag is unnecessary and may even cause different behavior depending on the WDL implementation used to run the code.